requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0

# Video processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Audio processing
SpeechRecognition>=3.10.0
pydub>=0.25.0

# Document generation
python-docx>=1.0.0
markdown>=3.5.0

# Utilities
tqdm>=4.66.0
